Fix failed tests

+++ tests/tests_resources/test_category.py
@@ -2,17 +2,30 @@ import json
 from tests.base import BaseTestCase
 from models.category import CategoryModel
 from models.recipe import RecipeModel
+from models.user import UserModel
 
 
 class CategoryTest(BaseTestCase):
     """Ensures that the category resource functions as expectected"""
+
+    def setUp(self):
+        super(CategoryTest, self).setUp()
+        with self.app() as client:
+            with self.app_context():
+                UserModel('testusername', 'testpassword').save_to_db()
+                auth_request = client.post('/login',
+                                           data=json.dumps(
+                                               {'username': 'testusername', 'password': 'testpassword'}),
+                                           headers={'Content-Type': 'application/json'})
+                auth_token = json.loads(auth_request.data)['access_token']
+                self.access_token = 'JWT {0}'.format(auth_token)
 
     def test_create_category(self):
         """Ensure that a recipe category is created """
 
         with self.app() as client:
             with self.app_context():
-                resp = client.post('/category/somerecipecategory')
+                resp = client.post('/category/somerecipecategory',headers={'Authorization': self.access_token})
 
                 self.assertEqual(resp.status_code, 201)
                 self.assertIsNotNone(
@@ -24,8 +37,8 @@ class CategoryTest(BaseTestCase):
         """Ensure that no duplicate categories are created"""
         with self.app() as client:
             with self.app_context():
-                client.post('/category/somerecipecategory')
-                resp = client.post('/category/somerecipecategory')
+                client.post('/category/somerecipecategory', headers={'Authorization': self.access_token})
+                resp = client.post('/category/somerecipecategory', headers={'Authorization': self.access_token})
 
                 self.assertEqual(resp.status_code, 400)
 
@@ -34,7 +47,7 @@ class CategoryTest(BaseTestCase):
         with self.app() as client:
             with self.app_context():
                 CategoryModel('somecategory').save_to_db()
-                resp = client.delete('/category/somecategory')
+                resp = client.delete('/category/somecategory', headers={'Authorization': self.access_token})
 
                 self.assertEqual(resp.status_code, 200)
                 self.assertDictEqual({'message': 'Category deleted'},
@@ -45,7 +58,7 @@ class CategoryTest(BaseTestCase):
         with self.app() as client:
             with self.app_context():
                 CategoryModel('somecategory').save_to_db()
-                resp = client.get('/category/somecategory')
+                resp = client.get('/category/somecategory', headers={'Authorization': self.access_token})
 
                 self.assertEqual(resp.status_code, 200)
                 self.assertDictEqual({'id': 1, 'name': 'somecategory', 'recipes': []},
@@ -55,7 +68,7 @@ class CategoryTest(BaseTestCase):
         """Test whether a category cannot be found"""
         with self.app() as client:
             with self.app_context():
-                resp = client.get('/category/somecategory')
+                resp = client.get('/category/somecategory', headers={'Authorization': self.access_token})
 
                 self.assertEqual(resp.status_code, 404)
                 self.assertDictEqual({'message': 'Category not found'},
@@ -69,7 +82,7 @@ class CategoryTest(BaseTestCase):
                 RecipeModel(
                     'African Tea', "Add two spoonfuls of tea leaves...", 1).save_to_db()
 
-                resp = client.get('/category/Beverage')
+                resp = client.get('/category/Beverage', headers={'Authorization': self.access_token})
                 self.assertEqual(resp.status_code, 200)
                 self.assertDictEqual({'id': 1, 'name': 'Beverage', 'recipes': [{'name': 'African Tea', 'description': 'Add two spoonfuls of tea leaves...'}]},
                                      json.loads(resp.data))
@@ -80,7 +93,7 @@ class CategoryTest(BaseTestCase):
             with self.app_context():
                 CategoryModel('somecategory').save_to_db()
 
-                resp = client.get('/categories')
+                resp = client.get('/categories', headers={'Authorization': self.access_token})
                 self.assertDictEqual({'categories': [{'id': 1, 'name': 'somecategory', 'recipes': []}]},
                                      json.loads(resp.data))
 
@@ -92,6 +105,6 @@ class CategoryTest(BaseTestCase):
                 RecipeModel(
                     'African Tea', 'Add two spoonfuls of tea leaves...', 1).save_to_db()
 
-                resp = client.get('/categories')
+                resp = client.get('/categories', headers={'Authorization': self.access_token})
                 self.assertDictEqual({'categories': [{'id': 1, 'name': 'Beverages', 'recipes': [{'name': 'African Tea', 'description': 'Add two spoonfuls of tea leaves...'}]}]},
                                      json.loads(resp.data))
